pyproject.toml
[project]
name = "githubctl"
version = "0.1.0"
description = ""
authors = [
    {name = "CarlosLeonSalazar",email = "165884029+CarlosLeonSalazar@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer[all] (>=0.16.0,<0.17.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "jmespath (>=1.0.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "githubctl", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
